extracted downloadHistory to utils
